Moving cert.rs into cert module, and splitting off extensions into ext.rs
